backend/requirements.txt
# Django 핵심
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework_simplejwt==5.5.0
gunicorn>=21.2
django-filter>=23.1

# DB 및 설정
psycopg2-binary==2.9.10
python-dotenv==1.1.0
python-decouple==3.8
sqlparse==0.5.3

# AWS S3 presigned URL
boto3==1.38.15
botocore==1.38.15
s3transfer==0.12.0

# 쿠폰/QR
qrcode==8.2
Pillow==11.1.0

# 응답 포맷, JWT 등
PyJWT==2.9.0
requests==2.32.3

# 시간, 날짜 관련
python-dateutil==2.9.0.post0
tzdata==2025.2
